Move `mesolve` to a specific folder (#33)
* Move mesolve to separate folder and split rouchon

* Move Rouchon solver options to mesolve folder

* Remove order argument from ME Rouchon solver

* Rename mesolve solver options to mesolver

* Fix @gautierronan comments

* Fix rebase
