+++ torchqdynamics/__init__.py
@@ -1,4 +1,5 @@
 from .utils import *
 from .mesolve import mesolve
+from .solver import *
 from .ssolve import ssolve
 from .smesolve import smesolve
